fix: changelog argument formatting in release workflow
Fix the formatting of the argument for the Changelog Action in the
release workflow. Removed an unnecessary line break within the `args`
value by removing the extra newline after `${{
needs.release.outputs.current_version`. This change ensures that the
arguments are correctly passed to the Changelog Action and the Github
release generation step during the workflow execution.